Apply lazy import for S3/GCS/Azure blob providers
