+++ neuro-sdk/tests/test_retry.py
@@ -28,7 +28,7 @@ async def test_first_fails(caplog: Any) -> None:
                 raise aiohttp.ClientError("Ouch!")
 
     assert count == 2
-    record = ("neuro_sdk.utils", logging.INFO, "Fails: Ouch!.  Retry...")
+    record = ("neuro_sdk", logging.INFO, "Fails: Ouch!.  Retry...")
     assert caplog.record_tuples == [record]
 
 
@@ -42,7 +42,7 @@ async def test_two_fail(caplog: Any) -> None:
                 raise aiohttp.ClientError("Ouch!")
 
     assert count == 3
-    record = ("neuro_sdk.utils", logging.INFO, "Fails: Ouch!.  Retry...")
+    record = ("neuro_sdk", logging.INFO, "Fails: Ouch!.  Retry...")
     assert caplog.record_tuples == [record] * 2
 
 
@@ -56,7 +56,7 @@ async def test_all_fail(caplog: Any) -> None:
                 raise aiohttp.ClientError("Ouch!")
 
     assert count == 3
-    record = ("neuro_sdk.utils", logging.INFO, "Fails: Ouch!.  Retry...")
+    record = ("neuro_sdk", logging.INFO, "Fails: Ouch!.  Retry...")
     assert caplog.record_tuples == [record] * 2
 
 
@@ -72,7 +72,7 @@ async def test_reset(caplog: Any) -> None:
                 raise aiohttp.ClientError("Ouch!")
 
     assert count == 7
-    record = ("neuro_sdk.utils", logging.INFO, "Fails: Ouch!.  Retry...")
+    record = ("neuro_sdk", logging.INFO, "Fails: Ouch!.  Retry...")
     assert caplog.record_tuples == [record] * 6
